requirements.txt
python-binance==1.0.17
python-telegram-bot==20.7
pandas==2.2.1
ta==0.10.2
python-dotenv==1.0.1
schedule==1.2.0
pytest==8.1.1
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
alembic==1.13.1
plotly==5.19.0
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.27.0